Split tmxutil.py into a proper python package
