+++ src/tmxutil/__init__.py
@@ -1,0 +1,5 @@
+try:
+	import pkg_resources
+	__version__ = pkg_resources.require('tmxutil-pkg-jelmervdl')[0].version
+except:
+	__version__ = 'dev'
