+++ src/tmxutil/__main__.py
@@ -1,0 +1,4 @@
+from .cli import entrypoint
+
+if __name__ == '__main__':
+	entrypoint()
